동일한 파일등록 로직을 메소드로 분리

+++ src/main/java/com/board/summernote/Service/BoardService.java
@@ -10,7 +10,9 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
@@ -20,8 +22,9 @@ import java.util.UUID;
 public class BoardService {
 
     private final BoardMapper boardMapper;
+    private String UPLOAD_PATH = "C:\\images\\";
 
-    public String ListPage(BoardDTO boardDTO, Model model) {
+    public String ListPage(Model model) {
         model.addAttribute("getList", boardMapper.findAll());
         return "list";
     }
@@ -41,54 +44,21 @@ public class BoardService {
         return "insertForm";
     }
 
-    public void BoardInsert(BoardDTO boardDTO, MultipartHttpServletRequest request, FileDTO fileDTO) {
+    public String BoardInsert(BoardDTO boardDTO, MultipartHttpServletRequest request, FileDTO fileDTO) {
         List<MultipartFile> files = request.getFiles("file");
-        String path = "C:\\images\\";
+        List ordList = new ArrayList<>();
+
+        for (int i = 1; i <= files.size(); i++) {
+            ordList.add(i);
+        }
 
         boardMapper.insertBoard(boardDTO);
 
-        if (files != null && !files.isEmpty()) {
-            int i = 1;
-            for (MultipartFile mf : files) {
-                String orgname = mf.getOriginalFilename();
-                long fileSize = mf.getSize();
-                String extension = orgname.substring(orgname.lastIndexOf(".") + 1);
-                UUID uuid = UUID.randomUUID();
-                orgname = orgname.substring(0, orgname.indexOf("."));
-                fileDTO.setFilename(orgname);
-                fileDTO.setPath(path);
-                fileDTO.setFilesize(fileSize);
-                fileDTO.setOrd(i);
-                fileDTO.setExt(extension);
-                fileDTO.setUuid(String.valueOf(uuid));
-                fileDTO.setRefType("notice");
-                String saveFile = path + uuid + "." + extension;
 
-                try {
-                    if (i > 5) {
-                        i = -1;
-                        break;
-                    } else {
-                        File file = new File(saveFile);
-                        file.mkdir();
-                        mf.transferTo(file);
-                        System.out.println("----------SaveFile Info----------");
-                        System.out.println("filename : " + orgname);
-                        System.out.println("filesize : " + fileSize + " B");
-                        System.out.println("filepath : " + path);
-                        System.out.println("fileorder : " + i);
-                        System.out.println("fileExt : " + extension);
-                        System.out.println("uuid : " + uuid);
-                        boardMapper.insertFile(fileDTO);
-                    }
-                    i++;
-                } catch (IllegalStateException e) {
-                    e.printStackTrace();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
+
+
+        return insertFile(files, fileDTO, boardDTO, ordList);
+
     }
 
     public void BoardDelete(BoardDTO boardDTO) {
@@ -96,12 +66,13 @@ public class BoardService {
         boardMapper.removeByIdx(boardDTO);
     }
 
-    public void BoardUpdate(BoardDTO boardDTO, MultipartHttpServletRequest request, FileDTO fileDTO, List<Integer> ordList) {
-        boardMapper.deleteFile(boardDTO); //reset File
+    public String BoardUpdate(BoardDTO boardDTO, MultipartHttpServletRequest request, FileDTO fileDTO, List<Integer> ordList) {
+        boardMapper.updateByIdx(boardDTO);
 
         List<MultipartFile> requestFiles = request.getFiles("fileList");
-        String path = "C:\\images\\";
-
+        if (requestFiles != null && !requestFiles.isEmpty()){
+            boardMapper.deleteFile(boardDTO); //reset File
+        }
         int fileindex = 0;
 
         List<String> filenameList = fileDTO.getFilenameList();
@@ -111,12 +82,11 @@ public class BoardService {
 
         //기존파일 Update처리
         for (int i = 0; i < filenameList.size(); i++) {
-            System.out.println(i);
             fileDTO.setFilename(filenameList.get(i));
             fileDTO.setFilesize(Long.parseLong(filesizeList.get(i)));
             fileDTO.setUuid(uuidList.get(i));
             fileDTO.setExt(extList.get(i));
-            fileDTO.setPath(path);
+            fileDTO.setPath(UPLOAD_PATH);
             fileDTO.setOrd(ordList.get(fileindex));
             fileDTO.setRefIdx(boardDTO.getIdx());
             fileDTO.setRefType("notice");
@@ -124,41 +94,64 @@ public class BoardService {
             fileindex++;
         }
 
-        // 신규파일 등록 처리
-        for(MultipartFile mf : requestFiles) {
-            System.out.println("----added file----");
-            String orgname = mf.getOriginalFilename();
-            long fileSize = mf.getSize();
-            String extension = orgname.substring(orgname.lastIndexOf(".") + 1);
-            UUID uuid = UUID.randomUUID();
-            orgname = orgname.substring(0, orgname.indexOf("."));
-            fileDTO.setFilename(orgname);
-            fileDTO.setPath(path);
-            fileDTO.setFilesize(fileSize);
-            fileDTO.setExt(extension);
-            fileDTO.setUuid(String.valueOf(uuid));
-            fileDTO.setOrd(ordList.get(fileindex));
-            fileDTO.setRefIdx(boardDTO.getIdx());
-            fileDTO.setRefType("notice");
-
-            String saveFile = path + uuid + "." + extension;
-
-            try {
-                mf.transferTo(new File(saveFile));
-                boardMapper.insertFile(fileDTO);
-                fileindex++;
-            } catch (IllegalStateException e) {
-
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-        boardMapper.updateByIdx(boardDTO);
-
+        //신규파일 Update처리
+        return insertFile(requestFiles, fileDTO, boardDTO, ordList);
     }
 
     public List<FileDTO> SelectFileList(BoardDTO boardDTO) {
        return boardMapper.findFile(boardDTO);
     }
 
+    public String insertFile(List<MultipartFile> requestFiles, FileDTO fileDTO, BoardDTO boardDTO, List<Integer> ordList) {
+
+        String data = "";
+
+        if (requestFiles != null && !requestFiles.isEmpty()) {
+            int fileindex = 0;
+            if (fileDTO.getFilenameList() != null && !fileDTO.getFilenameList().isEmpty()) { //기존 파일이 있는 지 체크
+                fileindex = fileDTO.getFilenameList().size();
+            }
+            for (MultipartFile mf : requestFiles) {
+                String orgname = mf.getOriginalFilename();
+                long fileSize = mf.getSize();
+                String extension = orgname.substring(orgname.lastIndexOf(".") + 1);
+                UUID uuid = UUID.randomUUID();
+                orgname = orgname.substring(0, orgname.indexOf("."));
+                fileDTO.setFilename(orgname);
+                fileDTO.setPath(UPLOAD_PATH);
+                fileDTO.setFilesize(fileSize);
+                fileDTO.setExt(extension);
+                fileDTO.setUuid(String.valueOf(uuid));
+                fileDTO.setOrd(ordList.get(fileindex));
+                fileDTO.setRefIdx(boardDTO.getIdx());
+                fileDTO.setRefType("notice");
+
+                String saveFile = UPLOAD_PATH + uuid + "." + extension;
+
+                try {
+                    if (ordList.get(fileindex) > 5) {
+                        break;
+                    } else {
+                        mf.transferTo(new File(saveFile));
+                        System.out.println("----------SaveFile Info----------");
+                        System.out.println("filename : " + orgname);
+                        System.out.println("filesize : " + fileSize + " B");
+                        System.out.println("filepath : " + UPLOAD_PATH);
+                        System.out.println("fileorder : " + ordList.get(fileindex));
+                        System.out.println("fileExt : " + extension);
+                        System.out.println("uuid : " + uuid);
+                        boardMapper.insertFile(fileDTO);
+                        fileindex++;
+                    }
+
+                } catch (IllegalStateException e) {
+                    data = "request error";
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    data = "File Error";
+                }
+            }
+        }
+        return data;
+    }
 }
